Basket/requirements.txt
fastapi
uvicorn
pydantic
loguru
aiokafka
aio-pika
python-dotenv
redis